Move `lap` struct into its own file
