+++ tui/stopwatch/components/laps/laps.go
@@ -13,15 +13,6 @@ import (
 // MODEL
 // -----
 
-type lap struct {
-	start time.Time
-	end   time.Time
-}
-
-func (l lap) duration() time.Duration {
-	return l.end.Sub(l.start)
-}
-
 // Represents a list of lap times
 type Model struct {
 	laps []lap // The collection of lap times
@@ -47,6 +38,7 @@ func (m Model) Len() int {
 	return len(m.laps)
 }
 
+// Returns the last lap time in the list or a fallback time if the list is empty
 func (m Model) LastOr(fallback time.Time) time.Time {
 	if m.Len() == 0 {
 		return fallback
@@ -54,6 +46,7 @@ func (m Model) LastOr(fallback time.Time) time.Time {
 	return m.laps[m.Len()-1].end
 }
 
+// Returns the shortest lap time in the list
 func (m Model) Shortest() int {
 	shortest := 0
 	for i, lap := range m.laps {
@@ -64,6 +57,7 @@ func (m Model) Shortest() int {
 	return shortest
 }
 
+// Returns the longest lap time in the list
 func (m Model) Longest() int {
 	longest := 0
 	for i, lap := range m.laps {
@@ -74,6 +68,7 @@ func (m Model) Longest() int {
 	return longest
 }
 
+// Returns the total time of all lap times in the list
 func (m Model) sum(i int) time.Duration {
 	var total time.Duration
 	for _, lap := range m.laps[:i+1] {
